requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
flask-openapi3==2.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
nose2==0.12.0
pydantic==1.10.2
pyrsistent==0.18.1
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
typing_extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.1
PyJWT==2.8.0
flask-bcrypt==1.0.1
